refactor: renamed context providers and moved styling to the top
