+++ frontend/main/src/components/context-provider/NotificationProvider.tsx
@@ -1,68 +1,138 @@
-import React, { useState, createContext } from 'react'
-import { Alert, AlertColor, Snackbar } from '@mui/material'
-
-interface FormProperties {
-    isOpen: boolean,
-    isSuccess: boolean,
-    message: string,
-    statusCode?: number | undefined
-  }
+import { HubConnectionBuilder } from '@microsoft/signalr'
+import React, { createContext, useContext, useEffect, useState, useMemo } from 'react'
+import { UserContext } from './UserProvider'
+import { NotificationModel } from '../../api/Client'
+import { _apiClient } from '../../App'
+import { ToastContext } from './ToastProvider'
+import { Button, Dialog, DialogActions, DialogContent, DialogContentText } from '@mui/material'
 
 interface NotificationProviderProps {
     children: React.ReactNode
   }
-
-export const NotificationContext =
-    createContext({
-        openNotification: (isSuccess: boolean, message: string) => {}
+  
+export const NotificationContext = createContext<{notifications: NotificationModel[],
+                                               unReadNotifications: number, 
+                                               setNotificationOpen: React.Dispatch<React.SetStateAction<NotificationModel | undefined>>
+                                               deleteNotification: (notification: NotificationModel) => void
+                                               readNotification: (notification: NotificationModel) => void}>({
+        notifications: [],
+        unReadNotifications: 0,
+        setNotificationOpen: () => {},
+        deleteNotification: () => {},
+        readNotification: () => {}
     })
 
 const NotificationProvider = (props: NotificationProviderProps) => {
-    const [ notification, setNotification ] = useState<FormProperties>({
-        isOpen: false,
-        isSuccess: false,
-        message: ''
-    })
-
-    const openNotification = (isSuccess: boolean, message: string) => {
-        setNotification({
-            isOpen: true,
-            isSuccess: isSuccess,
-            message: message
+    const [ notifications, setNotifications ] = useState<NotificationModel[]>([])
+    const [ notificationOpen, setNotificationOpen ] = useState<NotificationModel>()
+    const { user, token } = useContext(UserContext)
+    const { openToast } = useContext(ToastContext) 
+    const unReadNotifications = useMemo(() => {
+        let readCount: number = 0
+        notifications.forEach(n => {
+            if (!n.read) {
+                readCount ++
+            }
         })
+        return readCount
+    }, [notifications])
+
+    useEffect(() => {
+        const subscribeToNotifications = async () => {
+            try {
+                getNotifications()
+                const conn = new HubConnectionBuilder()
+                                 .withUrl('https://localhost:7208/NotificationHub')
+                                 .build()
+    
+                await conn.on('newMessage', notification => {
+                    getNotifications()
+                })
+                await conn.start().then(() => {
+                    conn.invoke('GetConnectionId', token).then((identifier) => {
+                        console.log(identifier)
+                    })
+                })
+            }
+            catch{
+                console.log('error!')
+            }
+        }
+        const getNotifications = async() => {
+            try {
+                const pastNotifications = await _apiClient.notificationsGET(user?.email)
+                setNotifications(pastNotifications)
+            }
+            catch(err: any) {
+                if (err.status !== 404) {
+                    openToast(false, 'Failed to load notifications! ' + err.message)
+                }
+            }
+        }
+
+        if (token && user) {
+            subscribeToNotifications()
+        }
+    }, [user, token])
+
+    const deleteNotification = async (notification : NotificationModel) => {
+        let newNotifications = [...notifications]
+        try {
+            const ind = newNotifications.indexOf(notification)
+            if (ind !== -1) {
+                await _apiClient.notificationDELETE(notification.id)
+                newNotifications.splice(ind, 1) 
+            }
+            setNotifications(newNotifications)
+            openToast(true, 'Successfully Deleted Notification')
+        }
+        catch(err: any) {
+            openToast(false, `Error Deleting Notification!: ${err.message}`)
+        }
     }
 
-    const handleClose = () => {
-        setNotification({
-          ...notification,
-          isOpen: false
-        })
-      }
-    
-      const handleSeverity = () => {
-        if(notification.isSuccess) {
-          return 'success' as AlertColor
+    const readNotification = async (notification : NotificationModel) => {
+        if (notification.read) {
+            return
         }
-        else {
-          return 'error' as AlertColor
+        let newNotifications = notifications
+        try {
+            const ind = newNotifications.indexOf(notification)
+            let newNotification = notification
+            newNotification.read = true
+            if (ind !== -1) {
+                await _apiClient.notificationPUT(newNotification)
+                newNotifications[ind] = newNotification 
+            }
+            setNotifications(newNotifications)
         }
-      }
+        catch(err: any) {
+            openToast(false, `Error Modifying Notification!: ${err.message}`)
+        }
+    }
+
+    const handleNotificationClose = () => {
+        setNotificationOpen(undefined)
+    }
 
     return (
-        <NotificationContext.Provider value={{openNotification}}>
-            {props.children}
-            <Snackbar
-            anchorOrigin={{vertical: 'bottom', horizontal: 'center'}} 
-            open={notification.isOpen} 
-            autoHideDuration={6000} 
-            onClose={handleClose}
-            key={'bottomcenter'}
+        <NotificationContext.Provider value={{notifications, unReadNotifications, setNotificationOpen, deleteNotification, readNotification}}>
+            { props.children }
+            <Dialog
+            open={notificationOpen !== undefined}
+            onClose={handleNotificationClose}
             >
-                <Alert severity={handleSeverity()} sx={{ width: '100%' }}>
-                {notification.message}
-                </Alert>
-            </Snackbar>
+                <DialogContent>
+                    <DialogContentText>
+                        {notificationOpen?.body}
+                    </DialogContentText>
+                </DialogContent>
+                <DialogActions>
+                    <Button onClick={handleNotificationClose}> Close </Button>
+                </DialogActions>
+            </Dialog>
         </NotificationContext.Provider>
     )
-}
-export default NotificationProvider
+}    
+
+export default NotificationProvider
