+++ frontend/main/src/components/context-provider/UserProvider.tsx
@@ -4,7 +4,7 @@ import { _apiClient } from '../../App'
 import { Box, Button, Modal, Typography } from '@mui/material'
 import { useNavigate } from 'react-router'
 import { Paths } from '../../utils/Constants'
-import { NotificationContext } from './NotificationProvider'
+import { ToastContext } from './ToastProvider'
 import { makeStyles } from 'tss-react/mui'
 
 const useStyles = makeStyles()(
@@ -45,7 +45,7 @@ const UserProvider = (props: UserProviderProps) => {
     const [ jwt, setJwt ] = useState<JwtModel>()
     const [ popup, setPopup ] = useState<boolean>(false)
 
-    const notificationContext = useContext(NotificationContext)
+    const toastContext = useContext(ToastContext)
     const navigate = useNavigate()
 
     useEffect(() => {
@@ -80,7 +80,7 @@ const UserProvider = (props: UserProviderProps) => {
             localStorage.removeItem('token')
             setPopup(false)
             navigate(Paths['Login'], { replace: true })
-            notificationContext.openNotification(false, 'Logged out due to inactivity')
+            toastContext.openToast(false, 'Logged out due to inactivity')
             setToken(null)
           }
         }, 30000)
@@ -89,7 +89,7 @@ const UserProvider = (props: UserProviderProps) => {
           clearInterval(interval)
         }
       }
-    }, [ jwt, navigate, notificationContext ])
+    }, [ jwt, navigate, toastContext ])
 
     const { classes } = useStyles()
 
@@ -103,7 +103,7 @@ const UserProvider = (props: UserProviderProps) => {
       setToken(null)
       setPopup(false)
       navigate(Paths['Login'], { replace: true })
-      notificationContext.openNotification(true, 'Successfully logged out')
+      toastContext.openToast(true, 'Successfully logged out')
     }
 
     const handleContinueSession = async () => {
@@ -112,10 +112,10 @@ const UserProvider = (props: UserProviderProps) => {
         localStorage.setItem('token', response.token ?? '')
         setToken(localStorage.getItem('token'))
         setPopup(false)
-        notificationContext.openNotification(true, 'Successfully refreshed user session')
+        toastContext.openToast(true, 'Successfully refreshed user session')
       }
       catch(err: any) {
-        notificationContext.openNotification(false, 'Failed to refresh session')
+        toastContext.openToast(false, 'Failed to refresh session')
       }
     }
